requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pillow==10.0.0
gitpython==3.1.40
requests==2.31.0
pymongo==4.6.0
python-dotenv==1.0.0